@findepi 2021-01-12
I believe this has been replaced with https://github.com/trinodb/trino-python-client/pull/60, so let me close  this.
Please reopen if i missed anything.